requirements.txt
apturl==0.5.2
asn1crypto==0.24.0
barcode==0.6.0
base64image==0.5.1
Brlapi==0.6.6
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
cryptography==2.1.4
cssselect2==0.2.1
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==2.1.3
django-cors-headers==2.4.0
djangorestframework==3.9.0
docopt==0.6.2
docutils==0.14
fpdf==1.7.2
httplib2==0.9.2
idna==2.6
image==1.5.27
img2pdf==0.3.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
Levenshtein-search==1.4.4
louis==3.5.0
lxml==4.4.1
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.0
meld==3.18.0
netifaces==0.10.4
numpy==1.17.0
oauth==1.0.1
olefile==0.45.1
opencv-python==4.1.2.30
pandas==0.23.4
pdf2image==1.10.0
pdf2jpg==1.0
pexpect==4.2.1
Pillow==6.1.0
protobuf==3.0.0
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyMySQL==0.9.2
PyNaCl==1.1.2
pyPdf==1.13
PyPDF2==1.26.0
pyRFC3339==1.0
python-apt==1.6.4
python-barcode==0.10.0
python-dateutil==2.8.0
python-debian==0.1.32
python-Levenshtein==0.12.0
pytz==2019.2
pyusb==1.0.2
pyxdg==0.25
PyYAML==3.12
pyzbar==0.1.8
reportlab==3.4.0
requests==2.20.1
requests-unixsocket==0.1.5
SecretStorage==2.3.1
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.0
ssh-import-id==5.7
svg2rlg==0.3
svglib==0.9.2
system-service==0.3
systemd-python==234
tinycss2==1.0.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
utl==1.0.1
virtualenv==16.7.2
wadllib==1.3.2
webencodings==0.5.1
xkit==0.0.0
xlrd==1.2.0
XlsxWriter==1.1.2
zope.interface==4.3.2